refactor : likes에서 like로 변경
