+++ src/main/java/com/sch/rezero/mapper/community/LikeMapper.java
@@ -1,6 +1,6 @@
 package com.sch.rezero.mapper.community;
 
-import com.sch.rezero.dto.community.likes.LikeResponse;
+import com.sch.rezero.dto.community.like.LikeResponse;
 import com.sch.rezero.entity.community.Like;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
